Add code for analyzing sample names
Analyzing sample data (both scraped and from gradedata.js) will be
helpful for building the best possible name-matching code.

I decided to add the name-analyzing code as a series of test cases
tucked away in a 'test' directory to emphasize that this code is
not part of the final product and is meant to assist development.